+++ cljs-test/igraph/browser.cljs
@@ -1,0 +1,5 @@
+(ns igraph.browser
+  (:require [doo.runner :refer-macros [doo-tests]]
+            [igraph.graph-test]))
+
+(doo-tests 'igraph.graph-test)